requirements.txt
pyYaml
pandas
numpy
nltk
matplotlib.pyplot
seaborn 
sklearn